+++ core/drivers/clk/clk_dt.c
@@ -8,6 +8,7 @@
 #include <drivers/clk_dt.h>
 #include <initcall.h>
 #include <kernel/boot.h>
+#include <kernel/dt.h>
 #include <kernel/dt_driver.h>
 #include <kernel/panic.h>
 #include <libfdt.h>
